Just use the home-relative gopath
This still doesn't work on Windows, but I'm starting to think that's not due to the path being wrong or anything like that, I think it's just a quirk of the Windows runner and it might not be something that can be worked around. And if it can be, then that's a separate task.

Signed-off-by: Yuri Norwood <106889957+norwd@users.noreply.github.com>